ENH: Add keyboard shortcuts for undo/redo
